requirements.txt
accelerate==1.10.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
diffusers==0.35.1
filelock==3.13.1
fsspec==2024.6.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.12.0.88
packaging==25.0
pillow==11.0.0
psutil==7.0.0
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
sympy==1.13.3
tokenizers==0.22.0
torch==2.8.0+cu128
torchvision==0.23.0+cu128
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.15.0
urllib3==2.5.0
zipp==3.23.0
